Move Tooltips, Tours components into directories
