--- assets/javascripts/kitten/components/tooltips/question-mark-with-tooltip.js
@@ -1,4 +1,0 @@
-import { withTooltip } from '../../hoc/with-tooltip'
-import { ButtonQuestionMarkIcon } from '../../components/buttons/button-question-mark-icon'
-
-export const QuestionMarkWithTooltip = withTooltip(ButtonQuestionMarkIcon, {})
