+++ assets/javascripts/kitten/components/tooltips/question-mark-with-tooltip/index.js
@@ -1,0 +1,4 @@
+import { withTooltip } from '../../../hoc/with-tooltip'
+import { ButtonQuestionMarkIcon } from '../../../components/buttons/button-question-mark-icon'
+
+export const QuestionMarkWithTooltip = withTooltip(ButtonQuestionMarkIcon, {})
