+++ assets/javascripts/kitten/components/tooltips/static-tooltip/index.js
