+++ assets/javascripts/kitten/components/tours/header-tour/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import { Tour } from '../../components/tours/tour'
+import { Tour } from '../../../components/tours/tour'
 
 // We directly use svg-react loader for the moment.
 // TODO: update webpack config to manage svg in jsx files.
-import { MixIllustration } from '../../components/illustrations/mix-illustration'
-import { LoudspeakerIllustration } from '../../components/illustrations/loudspeaker-illustration'
+import { MixIllustration } from '../../../components/illustrations/mix-illustration'
+import { LoudspeakerIllustration } from '../../../components/illustrations/loudspeaker-illustration'
 
 export class HeaderTour extends React.Component {
   constructor(props) {
