+++ assets/javascripts/kitten/components/tours/tour-popover/index.js
